Refactor for perf
